+++ src/gestures.tsx
@@ -18,7 +18,7 @@ const { width: WINDOW_WIDTH } = Dimensions.get('window');
 
 type ToastSwipeHandlerProps = Pick<
   ToastProps,
-  'important' | 'index' | 'numberOfToasts'
+  'important'
 > & {
   onRemove: () => void;
   style?: ViewStyle | (ViewStyle | undefined)[];
@@ -43,13 +43,10 @@ export const ToastSwipeHandler: React.FC<
   important,
   position: positionProps,
   onPress,
-  index,
-  numberOfToasts,
 }) => {
   const translate = useSharedValue(0);
   const {
     swipeToDismissDirection: direction,
-    gap,
     position: positionCtx,
   } = useToastContext();
   const position = positionProps || positionCtx;
@@ -103,6 +100,7 @@ export const ToastSwipeHandler: React.FC<
           translate.value = withTiming(0, {
             easing: Easing.elastic(0.8),
           });
+          return;
         }
 
         if (shouldDismiss) {
@@ -201,9 +199,10 @@ export const ToastSwipeHandler: React.FC<
             ? undefined
             : {
                 justifyContent: 'center',
-                marginBottom: gap,
               },
-          { width: '100%', zIndex: -(numberOfToasts - index) },
+          {
+            width: '100%',
+          },
           Platform.OS === 'android'
             ? {
                 opacity: 1,